Add the `take_over` mode for filestream inputs (#34292)
If a `filestream` input has the configuration parameter `take_over`
set to `true`, every `loginput` state record (in the registry) with
the `source` that matches at least one of the `filestream`'s
paths/globs will be taken over by this `filestream` input.

This means the existing `loginput` state entry gets converted into a
`filestream` entry (the `loginput` entry gets deleted).

The purpose of this mode is to make migration from `loginput` to
`filestream` as simple and smooth as possible by adding `take_over:
true` to the new `filestream` configuration. All offsets for input
files will be preserved and the `filestream` will continue ingesting
the files at the same point where the `loginput` stopped. This solves
the previously occurring data duplication (file re-ingestion) problem.