packetbeat/{decoder,protos/...}: remove logp.Recover uses and tighten recovers
It is generally not a good idea to recover from arbitrary panics and continue
operating. Unhandled panic are recovered and logged with termination at the beats
level.

In a couple of cases, the plugin uses panic/recover for flow control, this is
retained, but tightened up so that original panics are logged and incidental
panics that could leave the program in an invalid state are re-panicked.