Make `Host info` IPs valid
The `Host info` log entry is ingested to Elasticsearch by the
monitoring toolkit when running under Elastic Agent.

The monitoring events are ingested into a data
stream created out of the `logs` index template which is partially
defined by the `data-streams-mappings` component template.

The `data-streams-mappings` component template defines auto-mapping
for every field named `ip` to have the `ip` field type which does
not tolerate having a netmask as the IP address suffix.

This results in dropping events created out of this `Host info`
log entry.

This is to trim the netmasks and make sure that IP addresses we're
sending are valid, so the event is not dropped.