Restart Beat when certificates or key files are modified (#34416)
This commit enables any Beat to reexec when a output CA certificate file is modified. The main use case we're covering here is enabling "CA certificates reload" when running Metricbeat on ECK to monitor Elasticsearch. At the moment this feature is only supported on Linux and MacOS.

libbeat/management/management.go is also refactored and some unused bits of code are removed.