winlogbeat/docs: adjust XPath query term limit warning
The page linked to in the docs is now a 404 suggeesting that the issue
has been fixed. The only currently noted limitations on XPath are
described at https://learn.microsoft.com/en-us/windows/win32/wes/consuming-events#xpath-10-limitations

This page is captured by the wayback machine.

Windows Event Log supports a subset of XPath 1.0. The primary restriction is
that only XML elements that represent events can be selected by an event
selector. An XPath query that does not select an event is not valid. All
valid selector paths start with * or "Event". All location paths operate on
the event nodes and are composed of a series of steps. Each step is a
structure of three parts: the axis, node test, and predicate. For more
information about these parts and about XPath 1.0, see XML Path Language
(XPath). Windows Event Log places the following restrictions on the
expression:

* Axis: Only the Child (default) and Attribute (and its shorthand "@") axis
  are supported.
* Node Tests: Only node names and NCName tests are supported. The "*"
  character, which selects any character, is supported.
* Predicates: Any valid XPath expression is acceptable if the location paths
  conform to the following restrictions:
    * Standard operators OR, AND, =, !=, <=, <, >=, >, and parentheses are
      supported.
    * Generating a string value for a node name is not supported.
    * Evaluation in reverse order is not supported.
    * Node sets are not supported.
    * Namespace scoping is not supported.
    * Namespace, processing, and comment nodes are not supported.
    * Context size is not supported.
    * Variable bindings are not supported.
    * The position function, and its shorthand array reference, is supported
      (on leaf nodes only).
    * The Band function is supported. The function performs a bitwise AND for
      two integer number arguments. If the result of the bitwise AND is
      nonzero, the function evaluates to true; otherwise, the function
      evaluates to false.
    * The timediff function is supported. The function computes the difference
      between the second argument and the first argument. One of the
      arguments must be a literal number. The arguments must use FILETIME
      representation. The result is the number of milliseconds between the
      two times. The result is positive if the second argument represents a
      later time; otherwise, it is negative. When the second argument is not
      provided, the current system time is used.