Stop re-using processors defined in the config (#34761)
* Stop re-using processors defined in the config

After introducing the `SafeProcessor` wrapper in
https://github.com/elastic/beats/pull/34647 we started returning
errors when a processor is being used after its `Close` function has
been called.

This led to dropped events and error spam in logs but also confirmed
that the root cause of the problem was not just a race condition on
`Close` but re-used processors somewhere.

After a long investigation such code that's re-using processors was
finally found.

This is the change that removes re-using the processors and
instantiates them on each input restart.

* Fix linter issues

* Add changelog entry

(cherry picked from commit 5cfe62cb2f371e89f1af9e5e9c4f7144624d029c)

# Conflicts:
#	filebeat/channel/runner.go
#	libbeat/processors/safe_processor.go