libbeat/monitoring/inputmon - Ignore dataset metrics without name and id
This filters metrics from the /inputs/ endpoint for entities that lack an input `name` and `id`.

This is needed because we continued to use the "dataset" namespace in the monitoring
registry instead of using a new one like "inputs". Originally I planned to use to new namespace
but I stayed with `dataset` because I wanted to prevent a breaking change for existing users of the
/dataset API endpoint. This meant that metrics from the log input (and possibly others) that don't
yet follow the conventions need to be filtered.