fix grok patterns
The Elasticsearch made a behavioral change to the grok processor
(PR elastic/elasticsearch#92586), where in case a field matches more than
once, the processor will return a list of values. This change was
introduced in 8.7. Some of the filebeats pipelines were built over the
assumption that only the first match would be returned.

In the case of IIS, the `source.address` is repeated in the second part, by
which only the first match will be used, so, it's safe to remove the
second `source.address`

In the case of MySQL, according to their documentation:
 - the Schema is defined always at the
   beginning of the log, right after the "Thread_id:", so it's safe to
   remove the second `mysql.schema`
 - `thread_id` is more tricky, because it can be matched from different
   places, in this case, the potential matches are stored in 3 temporary
   fields, and then a new `script` processor does the job of using the
   correct temporary field and then removing it.