[Filebeat] Add support for dataset configs
This is a draft PR to show the approach that would be taken for implementation. Please review the approach not the code, as the code will change.

To make it easer for our users to ship to the [data stream naming scheme](https://www.elastic.co/blog/an-introduction-to-the-elastic-data-stream-naming-scheme) with filebeat, the same configs that are available in Elastic Agent should be added to Filebeat inputs:

```
data_stream.type
data_stream.dataset
data_stream.namespace
```

The idea is to hook into the shared inputConfig and add the 3 configuration variables there. If only one of the 3 values is set, the following defaults exist: `type: logs`, `dataset: generic`, `namespace: default`. As the data_stream config options overwrite the index config option, a check will be added that the two can't coexist. In addition, validation should be added to ensure the values set are valid.

Having these config options should also allow Elastic Agent eventually to just forward it to the inputs without having to set `index` value.

So far I have manually tested the change with `log` and `filestream` input.