packetbeat/beater: prevent double close of pb.done (#35324)
In some circumstances when a managed instance of packetbeat is restarted
Stop will panic with close of closed channel. This appears to be
happening because on reload the sniffer may have returned an error,
closing pb.done. If the packetbeat instance is then stopped, done is
closed a second time. So ensure that all closes are protected by
pb.stopOnce.

(cherry picked from commit 30e633c3514be8566b779ebfb6299b75f81fdde7)