x-pack/filebeat/input/cel: provide XML decoding functionality
Note a minor difference between the outcome of HTTPJSON XML parsing and
CEL XML parsing arrises due to the final result from CEL being passed
back to the input via JSON. This has the effect that numbers are coerced
to float, despite being type-hinted as integer. This should have minimal
impact, but it is important to be aware of during ingest.