x-pack/filebeat/input/aws-s3 - Fix test case race condition
Fixes race condition in tests reading/writing the `currentTime` func.

This guards access to the "clock" via a atomic.Value.
Another more elaborate fix would be to eliminate the global "clock" and inject
a clock into every instance via the constructor, but wiring that up to pass the
clock through to everywhere would be larger change. Given that this only affects
tests I'm using this simpler fix.

Fixes #35418