x-pack/filebeat/input/cel: add execution budget and relax CSV mime name (#35409)
This adds two features to improve user experience. The first is a guard
rail to prevent users unwittingly constructing a CEL program that enters
an infinite loop and blocking further processing. While CEL is not
Turing complete, there is a Turing escape hatch in the form of the
want_more field. If a user sets want_more to true and returns an event
the inner execution loop will be re-run. If an error in the CEL program
results in non-progress, this would result in an infinite loop. By
providing a bound on the number of executions each periodic invocation
can run, and warning when the budget is exceeded, this is mitigated.

The second is a change to the look-up table for MIME types. While the
documentation says that CSV MIME types are accessed via the name
"text/csv; header=absent" and "text/csv; header=present" (with a space),
it is probable that at some point someone will type this without a space
and get a failure to parse. So add the no-space variants to the table.