libbeat/monitoring/inputmon - Add defensive check for empty ID or input type
Add a defensive check to prevent metrics from being registered into the global
'dataset' namespace when the ID or input type are empty.

This is not in response to any known problem. There was already a similar check
on HTTP /inputs/ endpoint that would have prevented metrics from inputs with
an ID or input type from being exposed. This pushes that check earlier and in
effect "null routes" those metrics an a manner that is transparent to the callers.

I would consider an empty input type to be a developer mistake whereas an the
ID is user controllable. So you could conceive of having a hard error (like a panic)
if input type were empty. At the moment I'm taking a softer approach in the defense.

Relates #35354