Remove unnecessary logging (#35426)
* Remove unnecessary logging

The log line remove by this commit is unnecessary, it is not an actual
error in the sense something will not work. It only means the
harvester is already running for this file so another one does not
need to be started. Keeping this log line will only spam our log files
and make users worried.

* PR improvements

* Restoring the error context.