x-pack/filebeat/input/azureblobstorage: ensure checkpoint is valid after cursor filling
If the input is run and loads a cursor state with no partially processed
data, the CheckPoint.PartiallyProcessed map may be left nil. This
results in a nil-map assignment panic when updates are made to the
internal checkpoint. Currently the only other place PartiallyProcessed
is set is on CheckPoint value construction, so ensure that it is also
done here.