winlogbeat/sys/wineventlog: don't panic on nil iterator/renderer close (#35423)
In the case that an experimental API EventLog has been opened against a
non-available log channel, the EventLog will be closed. This
unconditionally closes the EventIterator and Renderer that the EventLog
hold. The EventIterator is only valid after a successful call to Open
and so may be nil when a valid EventLog is closed, resulting in a nil-deref
panic.

	panic: runtime error: invalid memory address or nil pointer dereference
	[signal 0xc0000005 code=0x0 addr=0x0 pc=0xc33574]

	goroutine 96 [running]:
	github.com/elastic/beats/v7/winlogbeat/sys/wineventlog.(*EventIterator).Close(0x0)
		github.com/elastic/beats/v7/winlogbeat/sys/wineventlog/iterator.go:199 +0x34
	github.com/elastic/beats/v7/winlogbeat/eventlog.(*winEventLogExp).Close(0xc00056cc80)
		github.com/elastic/beats/v7/winlogbeat/eventlog/wineventlog_experimental.go:347 +0xb6
	github.com/elastic/beats/v7/winlogbeat/beater.(*eventLogger).run.func3()
		github.com/elastic/beats/v7/winlogbeat/beater/eventlogger.go:128 +0x9f
	github.com/elastic/beats/v7/winlogbeat/beater.(*eventLogger).run(0xc0001aa600, 0xc0000b9020, {0x3608470?, 0xc0002bdb80}, {{0x0, 0x0}, 0x0, {0x0, 0x0, 0x0}, ...}, ...)
		github.com/elastic/beats/v7/winlogbeat/beater/eventlogger.go:149 +0xa6a
	github.com/elastic/beats/v7/winlogbeat/beater.(*Winlogbeat).processEventLog(0x0?, 0x0?, 0x0?, {{0x0, 0x0}, 0x0, {0x0, 0x0, 0x0}, {0x0, ...}}, ...)
		github.com/elastic/beats/v7/winlogbeat/beater/winlogbeat.go:203 +0xb3
	created by github.com/elastic/beats/v7/winlogbeat/beater.(*Winlogbeat).Run
		github.com/elastic/beats/v7/winlogbeat/beater/winlogbeat.go:171 +0x37a

So place a nil-check in the preamble to the EventIterator and Renderer's Close
methods and return an error to be logged in the case that they are nil. A call
to Renderer.Close should never encounter a nil receiver, but include the guard
for prudence.