auditbeat/module/file_integrity/fileorigin_test.go - Create file per sub-test
Instead of sharing the same file while handle is open across sub-tests, create
a new temp file for each sub-test and close it after creating it.

I couldn't reproduce the flaky test case, but I figured it can't hurt to further isolate
each sub-test with separate files. The tests are each modifying the file extended
attributes (so may be there is some race related to the filesystem extended attributes?).

Relates #33033