winlogbeat: allow termination while attempting to open channels (#35474)
Currently, if winlogbeat is attempting to open a channel that is not
available because it is not found or there is a recoverable error, it
will loop and retry. Since signals have been sent to notification
handlers by this point, cancellation of the loop is not possible without
agreement. So add a cancellation select at the head of the retry loop to
allow termination.