x-pack/filebeat/input/awss3: relax queue_url constraints for testing
Currently, the queue_url option is used to obtain the region name for
the SQS receiver. This prevents pointing the input at other sources for
testing, so add a region_name option to provide an alternative way to
provide the region name for non-AWS endpoints.

To avoid confusion, prevent using the region_name option in conjunction
with amazonaws.com endpoints and document this.