Enable dependabot for Go modules
This enables the GitHub dependabot which will open PRs
to keep the go.mod file updated when new releases are made
for the project's dependencies.

This configures it to watch the go.mod file and update it
when an Elastic project is released (github.com/elastic/*).
I wanted to keep the scope small to start with, but this
`allow` filter can be removed or expanded in the future.