Add post-dependabot GH action to update NOTICE.txt (#35538)
In elastic/beats, any changes to dependencies contained in go.mod requires the
change to be reflected in the NOTICE.txt file. When dependabot creates a
branch for a go_modules change this will update the NOTICE.txt file for that
change.

Relates #35525

Also add Team:Elastic-Agent-Data-Plane label to dependabot PRs.