packetbeat: fix handling of npcap.never_install configuration options from fleet (#35541)
Previously, the config option for blocking Npcap DLL installation on Windows under
fleet was not properly handled due to differences in the shape of the configs by
the time the decision is made for the action. This adds a tested helper to ensure
that fleet configurations are properly handled.

Also add some logging at the beginning of the install process and a note when a
previously installed Npcap is disabled.