filebeat/input/tcp: fix call to get unspecified IPv6 addresses (#35637)
The call incorrectly uses the IPv4 address list to populate the IPv6
unspecified address list. This will result in a panic when the more IPv6
addresses than IPv4 addresses appear in the the hosts IP list as returned
by net.LookupIP().

Also add system tests to cover this and the equivalent code in the UDP
input.