dependabot - Segment by responsible team
Segment the dependabot config based on the responsible GitHub
team in order to make it more clear who is responsible for approving
the change.

The idea is that the repsonsible team is in the best position to know
if the dependency should be updated, whether it should be backported,
and what should be listed in the changelog.