auditbeat/module/file_integrity: remove garbled PE executable test file
This file is detected by some malware detection systems as malicious
likely due to the source obfuscation that garble does to it. The testing
of the underlying library already tests the infrastucture, so removing
the test here does not harm coverage.

The executable itself is a garble built compilation of the file at
https://github.com/elastic/toutoumomoma/blob/dev/testdata/main.go. In
that repository the executable is built on the fly during testing rather
than being retained as an asset.