Add new `fingerprint` file identity
This is a new alternative to existing options like `native`, `path`
and `inode_marker`.

Unlike the existing options, this file identity does not rely on any
file system metadata and uses only the file size and its content.

Users can specify what amount of bytes is used to fingerprint the
beginning of each file.

This identity is supposed to be more stable and less affected by the
environment/setup of the users.