Draft input metrics about events and records
In this first iteration, the metrics keep track of the following data
types:

- events
- records

Events are the event delivered by the event hub; each event usually
contains a list of records.

Records are the actual logs from Azure; the input creates one document
in Elasticsearch for each record.

This draft keeps track of the following conditions.

Events:

- received: the input received from an event from the Event Hub
- sanitized: the event contains invalid JSON, and the input tried fixing it
- deserialization failed: the event contains invalid JSON; sanitization was ineffective
- processed: the input processed all the records in the event

Records:

- received: the input  unpacked a record from an event
- serialization failed: failed to serialize the record for dispatching
- processed: the input dispatched the record to the outlet