x-pack/filebeat/input/http_endpoint: make sure address is removed on server closure (#35903)
Previously the server address registry was not updated when the server was cancelled,
leaving invalid state in the registry and preventing reconfiguration of servers.
Ensure that the server address entry is deleted from the pool when the server is
cancelled.

In addition the pool lock was not being released in the case of a failed
TLS consistency check. The change here without fixing that resulted in a
deadlock. This is also fixed.

(cherry picked from commit 62979b558ccbd6b221dcb2268c5e73665ba653d9)