Add new `fingerprint` file identity (#35734)
This is a new alternative to existing options like `native`, `path`
and `inode_marker`.

Unlike the existing options, this file identity does not rely on any
file system metadata and uses only the file size and its content.

Users can specify what amount of bytes is used to fingerprint the
beginning of each file, optionally it's possible to set an offset from the beginning.

This identity is supposed to be more stable and less affected by the
environment/setup of the users.

This change also contains a few performance optimisations of how we work with the filesystem and watch for file changes.

(cherry picked from commit b701377c9b12d80bc666a147253395ada8376d67)

# Conflicts:
#	filebeat/input/filestream/config.go
#	filebeat/input/filestream/fswatch.go
#	filebeat/input/filestream/fswatch_test.go
#	filebeat/input/filestream/identifier.go
#	filebeat/input/filestream/prospector.go
#	filebeat/input/filestream/prospector_test.go
#	libbeat/metric/system/cgroup/cgcommon/metrics_test.go