Pin PyYAML version to 5.3.1 to avoid CI errors temporarily (#36091)
CPython 3.0 was released recently which appears to have introduced a regression that leads to failures when installing PyYAML (and perhaps other packages too). This is a temporary fix and this commits needs to be reverted when a proper fix is available.

(cherry picked from commit 46adbacfb5a89a02c025b11a66a6e5dbbaf4514a)

# Conflicts:
#	libbeat/tests/system/requirements.txt
#	libbeat/tests/system/requirements_aix.txt
#	metricbeat/Dockerfile