Fix empty new file edge case (#36076)
It's possible that file scanning happens when the file was created but
was not yet written to. In this case the size is 0.

We should not spawn any resources (e.g. harvesters) for such files until they actually
have some content. We create events only when read something from a
file, so having a harvester on an empty file is not useful.

Not handling this edge case also causes our tests to be flaky,
sometimes an expected size does not match an actual size (0).

(cherry picked from commit 0cd57755c1bc7cef14a94f71970982a4e19dcb5d)

# Conflicts:
#	filebeat/input/filestream/fswatch_test.go