x-pack/filebeat/input/cel: make now global static per evaluation (#36107)
Previously now was static for the life of the program, which
corresponded to the life of the input. This could lead to incorrect and
surprising times being provided when the global was used. Obtain a now
value before starting each evaluation and use it to shadow the CEL now
global (and share it with the Go logging to allow correlation between
these).

(cherry picked from commit b6c377c88447c3e922d678147c3c504d95200a3c)