Report status per unit
This commit updates the ManagerV2 to set status per Unit based on its
inputs. If any input on a Unit returns an error when starting the
whole Unit is set as failed. If multiple inputs return an error, all
errors are reported in the `Message` field.

If the output unit returns an error when starting, only the output
Unit is set as failed. All other input unit states are not
modified (this was the behaviour before this commit).