x-pack/winlogbeat/module/powershell: improve details regexp pattern performance (#36178)
The details regexp pattern for the powershell data stream was improved
in the windows integration in elastic/integrations#6154 to reduce
backtracking costs that caused regexp costs to exceed the set runtime
limits. The same pattern on other data streams and in winlogbeat was not
updated. This change brings winlogbeat up to date with the change.

(cherry picked from commit d8db41b7713019039d0c7bc132e3e96edc062712)