github: unify PR description and commit message
To improve the quality of commit messages in the revision history,
indicate that the commit message will be part of the PR review process
and integrate the WHAT and WHY questions in the change explanation.

Remove the checklist item for PR title and description as it becomes
redundant with this new text.