Changelog: adding 8.7.0 and 8.7.1 to release-notes.html (#36237)
<!-- Type of change
Please label this PR with one of the following labels, depending on the
scope of your change:
- Bug
- Enhancement
- Breaking change
- Deprecation
- Cleanup
- Docs
-->

## Proposed commit message

<!-- Mandatory
Explain here the changes you made on the PR.

Please explain:

- WHAT: patterns used, algorithms implemented, design architecture,
message processing, etc.
- WHY:  the rationale/motivation for the changes

This text will be pasted into the squash dialog when the change is
committed and will be
a long term historical record of the change to help future contributors
understand the
change, please help them by making it clear and comprehensive, they may
be you.

If the commit title is adequate to describe both of these things, The
text here may be omitted
or replaced with "See title". The title of the PR will be used as the
commit message title when
the merge is made and the "See title" marker will be removed if present.

The text here and the PR title will be subject to the PR review process.
-->

We forgot to modify `libbeat/docs/release.asciidoc` we added changelog
on the 8.8 branch.
The changelog content was already present.
 
## Checklist

<!-- Mandatory
Add a checklist of things that are required to be reviewed in order to
have the PR approved

List here all the items you have verified BEFORE sending this PR. Please
DO NOT remove any item, striking through those that do not apply. (Just
in case, strikethrough uses two tildes. ~~Scratch this.~~)
-->
- [x] I have made corresponding changes to the documentatio

## Related issues

<!-- Recommended
Link related issues below. Insert the issue link or reference after the
word "Closes" if merging this should automatically close it.

- Closes #123
- Relates #123
- Requires #123
- Superseds #123
-->
- Relates https://github.com/elastic/beats/issues/35580