x-pack/filebeat/input/cel: improve error reporting from Test run (#36271)
There appears to be a test flakiness in the cel package tests that cause the HTTP-
dependent tests to fail after the retry_failure test, with url is unreachable. To
help identify the cause of the failure, improve the error reporting from the Test
method which is the path that is failing.