dev-tools/packaging/templates/rpm: add warning when /etc/rc.d/init.d/functions is missing (#36188)
On RHEL9, the functions script is no longer present resulting in failure
when init-script based service management commands are used. We still
support at least one distribution that provides init-script based
service management, so we need to retain these scripts, but provide a
warning to users on systems where these will not work pointing to the
new approach.