filebeat/input/{tcp,udp}: fix resolution of device address during metric initialisation
The net package may return IPv4 addresses as an IPv6len net.IP value
with a ::ffff: prefix before the four IPv4 octets[1]. This means that we
may incorrectly assign all addresses to the IPv6 class rather than to
the IPv4 class that is expected. This results in loss of metrics for the
input and confusing warnings complaining about an absent address in
/proc/net/{tcp,udp}6.

Fix this by using the To4 and To16 methods which are guaranteed to
return a net.IP that is the expected length for the class and compare
against that length.

[1]https://cs.opensource.google/go/go/+/refs/tags/go1.21.0:src/net/ip.go;l=62