Make naming consistent with the main branch
In https://github.com/elastic/beats/pull/36065 a few fields were
renamed in order to clarify their purpose. Unfortunately, this rename
was a part of a new feature PR which was not supposed to be backported
to 7.17.

However, backporting some other changes related to this code
has now become challenging and results in build failures (note the
fixing commits):

- https://github.com/elastic/beats/pull/36095
- https://github.com/elastic/beats/pull/36264

This PR makes the naming consistent with the main branch, so we can
easily backport changes.