Add config param for bulk response filtering
This commit adds a new parameter in the output settings called
`bulk_response_filtering` which by default is going to be true.

This parameter when is true is going to append the parameter of the
ResponseFiltering to the Bulk request the beat is doing.
We are adding this parameter so that we can opt-out if we want to
diagnose a problem or if someone want to keep using the older way we
execute bulk requests.

Signed-off-by: Alexandros, Sapranidis <alexandros@elastic.co>