.github/dependabot.yml - Revert team based segmentation
Dependabot does not allow multiple configurations for a given
Go module. So this reverts the change to segment the configuration
along team boundaries with explicit team-based reviewer requirements.
It was complaining about

The property '#/updates/1' is a duplicate. Update configs must have a unique combination of 'package-ecosystem', 'directory', and 'target-branch'
The property '#/updates/2' is a duplicate. Update configs must have a unique combination of 'package-ecosystem', 'directory', and 'target-branch'

as pointed out in https://github.com/elastic/beats/pull/36158#issuecomment-1665700688.