auditbeat/module/file_integrity: encode POSIXACLAccess as []byte (#36351)
When encoded as a string, the logic within the module works correctly,
but the logged data is passed through the JSON serialisation logic which
will ensure that all the unicode code points are valid. It is not
necessarily true that all the bytes in the data for this field are valid
(almost guaranteed not to be given that invalid qualifiers are
represented by 0xffffffff).

So use a []byte to hold this data. This has the additional advantage
that the representation in the logs will be almost exactly what getfattr
shows for this; the only difference being a "0s" prefix added by
getfattr.

(cherry picked from commit 74368951c5b12b2f20de2dc367b59e0fcc8bc5cc)