libbeat/tests/compose: Fix conflict in container creation for integration tests (#36238)
Migrate from using ContainerKill to ContainerRemove as the former terminates the container process but does not remove the container from the docker host but the latter does the removal as well i.e., removes the container and associated links and volumes. Without this change, sometimes the creating containers results in error because of the existence of containers (orphans) resulting in a conflict.

Reduce the code smells by addressing the issues reported by golangci-lint and also update outdated doc comments.