golangci.yml: silence invalid linter complaints on use of logp.TestingSetup (#36474)
A logp.TestingSetup call can only fail when no logp.ToObserverOutput()
option has been provided and one of syslog, file or event log output has
been selected. There is no case in the code base where this is true.
So disable the errcheck linter for this function.

The only way that the function can return an error is when a user makes
use of a custom logp.Option function since none of the provided Options
alter the logp.Config to cause output to failable destinations.