winlogbeat/eventlog: ensure event loggers restart metric collection when handling recoverable errors
When winlogbeat's event loggers encounter recoverable errors they handle
this by closing and reopening the channel. This causes the metric
collection for the beat and dependent winlog filebeat input to lose
metric collection as metric registration only occurs on configuration.
So move metric registration to the channel open method. To avoid
reregistration of metrics, use the nilness of the metric field to guard
against this.